requirements.txt
# Python 3.11 or above
snowflake-connector-python
python-dotenv==1.0.1
openai==1.43.0
tiktoken==0.7.0
pytesseract
streamlit
pdf2image
faiss-cpu
cv2
PyMuPDF
fitz
sklearn
azure-storage-file-datalake
azure-functions
python-tabular
pyadomd
msal
numpy
pandas
pymsteams
azure-storage-blob
# llamaindex
llama-cloud==0.0.17
llama-index==0.11.8
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.8
llama-index-embeddings-azure-openai==0.2.5
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-azure-openai==0.2.1
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.5
# langchain
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
# haystack
haystack-ai==2.5.1
haystack-experimental==0.1.1
# dspy
dspy-ai==2.4.14
